Add logging in case of failed /tmp cleanup

+++ executor/app/src/compute_horde_executor/executor/management/commands/run_executor.py
@@ -716,11 +716,19 @@ class JobRunner:
             "sh",
             "-c",
             f"rm -rf {shlex.quote(root_for_remove.as_posix())}/*",
-            stdout=asyncio.subprocess.DEVNULL,
-            stderr=asyncio.subprocess.DEVNULL,
+            stdout=asyncio.subprocess.PIPE,
+            stderr=asyncio.subprocess.PIPE,
         )
-        await process.wait()
-        self.temp_dir.rmdir()
+        stdout, stderr = await process.communicate()
+        if process.returncode == 0:
+            self.temp_dir.rmdir()
+        else:
+            logger.error(
+                f"Failed to clean up {self.temp_dir.as_posix()}/: process exited with return code {process.returncode}\n"
+                "Stdout and stderr:\n"
+                f"{truncate(stdout.decode())}\n"
+                f"{truncate(stderr.decode())}\n"
+            )
         if self.is_streaming_job:
             process = await asyncio.create_subprocess_exec(
                 "docker", "network", "rm", self.job_network_name
